Changes log 6: - I added 1 more icon: - CrossIcon
- I added Modal component. Maybe I will improve it improve it later, but for now it seems working fine. I haven't tested it properly though. So I will do it later.
Also I added div with id="modal" to the Layout component to use it.

- I added DataLoader component. I will use it in the future during requests for the data about nannies from the FireDB

- I modified SortByDropDown component. Now it has three more props:
- optionsList - list of the dropDown items to fill it with
- onChange - method that is being called when selectedItem changes. It takes selectedItem value as a parameter and returns void.
- defaultOption - number that corresponds to the serial number of the element from the optionsList array that will be set by default as a selectedElement.

I changed the rest of the code in the SortByDropDown component accordingly to this new 3 props.
Also I made this dropDown more friendly to the keyboard. I wrapped text in the li items into buttons. And added tabIndex attribute. It is interesting how it works, so you can look at it again, by the way. Also I changed css styles, so that it looks nice now.

I modified code for the SortByDropDown in the Filters component accordingly to the 3 new props in the SortByDropDown component

- I added onClick prop and logic to the LoadMore button. You can look it in the LoadMoreButton component file

- I added Modal component to the NannyCard component.
Also I deleted fixed height from the NannyCard in the css

- I set pointer-events to none for the :after circle in the Navigation component

- I created Filter enum in the /src/types/data folder

- I created filters constant in the /src/constant folder. It cobverts Filter enum into an array of values from that enum

- I created filterNannies function in the /src/helpers folder for filtering data from the nannies list on the nannies page

- I cut code that I used for testing purposes from the Home page, modified it and moved to the Nannies page

- I started adding logic and state to the Nannies page. It is done for about 60%. For more details take a look at the NanniesPage file in the /src/pages/NanniesPage folder