Update for NanniesList in the NanniesPage
I think that I fixed the problem when the card didn' shrink after sorting filter change. I'm not sure if it is the right way to do it but for now it works fine. I just changed key value of the <li> tag during map from "data.name" to  "filter + data.name"