BIG update without any meaningful comment
This is a big update that was supposed to fix some bugs from the past and add new functionality to make the design more consistent with the technical specifications. I don’t know if I succeeded or not, but 1) the code looks disgusting and 2) so far everything is working. Most likely I will rewrite and clean it up in the future, but for now I just want to save it so that it will be easier for me to work and be able to roll back if something happens